+++ src/pl/io4/model/Cachable.java
@@ -1,0 +1,9 @@
+package pl.io4.model;
+
+/**
+ * Created by Kuba on 2017-04-25.
+ */
+public interface Cachable {
+	void cache();
+	void load();
+}
