+++ src/pl/io4/model/ProductsMachine.java
@@ -1,0 +1,15 @@
+package pl.io4.model;
+
+/**
+ * Created by Kuba on 2017-04-25.
+ */
+public class ProductsMachine implements Cachable {
+
+	public void cache() {
+
+	}
+
+	public void load() {
+
+	}
+}
