+++ src/pl/io4/model/ShopsMachine.java
@@ -1,0 +1,15 @@
+package pl.io4.model;
+
+/**
+ * Created by Kuba on 2017-04-25.
+ */
+public class ShopsMachine implements Cachable {
+
+	public void cache() {
+
+	}
+
+	public void load() {
+
+	}
+}
